MANY SMALL CHANGES, MAINLY TRY TO FIX VIEWING WINDOW ISSUE
The Viewing Window issue is that when the fitting polynomial
of the main graph goes to infinity (or just large numbers)
within the x range of the data, the window is very zoomed out.

Fix: Figure out a transformation to only show the y range of the
data +- 5% of the maximum difference of y values.